Extract about page logic into HasAboutUsCorporateInformationPage
I'm about to move the content for a WorldwideOrganisation from the
CorporateInformationPage about page onto the WorldwideOrganisation
itself. Once that move is completed, WorldwideOrganisation will still
need to support CorporateInformationPages, but not the about page
special case; whereas Organisation will still need to support the about
page special case. Extracting this concern will mean the change to
WorldwideOrganisation will be easier.
