+++ app/models/has_corporate_information_pages.rb
@@ -8,14 +8,6 @@ module HasCorporateInformationPages
     end
   end
 
-  def summary
-    about_us.summary if about_us.present?
-  end
-
-  def body
-    about_us.body if about_us.present?
-  end
-
   def unused_corporate_information_page_types
     CorporateInformationPageType.all - corporate_information_pages.map(&:corporate_information_page_type)
   end
@@ -28,13 +20,5 @@ module HasCorporateInformationPages
 
   def published_corporate_information_pages
     corporate_information_pages.published
-  end
-
-  def about_us
-    @about_us ||= corporate_information_pages.published.for_slug("about")
-  end
-
-  def draft_about_us
-    @draft_about_us ||= corporate_information_pages.draft.for_slug("about")
   end
 end
