+++ app/models/worldwide_organisation.rb
@@ -16,6 +16,7 @@ class WorldwideOrganisation < ApplicationRecord
   has_many :edition_worldwide_organisations, dependent: :destroy, inverse_of: :worldwide_organisation
   # This include is dependant on the above has_many
   include HasCorporateInformationPages
+  include HasAboutUsCorporateInformationPage
 
   has_many :editions, through: :edition_worldwide_organisations
